Revert "relayer: module chain::cosmos::test_utils"
This reverts commit 6a8e31f7d150f0596ddad73a87ef05a6ffd38af3.
